refactoring optionwindow_main.py file name
