+++ OptionWindow/test/test_optionwindow_main.py
@@ -2,7 +2,7 @@
 
 import sys
 from PyQt4 import QtGui
-from AQTrader.OptionWindow.optionwindow.main import MainForm
+from ..optionwindow.optionwindow_main import MainForm
 
 
 class TestClass(object):
